render2: fix overflow in window computation.

+++ SUPer/render2.py
@@ -211,7 +211,7 @@ class WindowOnBuffer:
         for sr in self.srs:
             mxy[:] = np.min([np.asarray((sr.y,  sr.x)),  mxy], axis=0)
             Mxy[:] = np.max([np.asarray((sr.y2, sr.x2)), Mxy], axis=0)
-        mxy, Mxy = np.uint32((mxy, Mxy))
+        mxy, Mxy = np.int16((mxy, Mxy))
         return Box(mxy[0], max(Mxy[0]-mxy[0], 8), mxy[1], max(Mxy[1]-mxy[1], 8))
 
     def area(self) -> int:
